requirements.txt

Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-filter>=23.0
python-dotenv>=1.0
Pillow>=10.0
